added realization
