--- Properties/AssemblyInfo.cs
@@ -1,9 +1,0 @@
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: ComVisible(false)]
-
-[assembly: Guid("680e8f73-2f7e-4f89-b3fa-8c9c2e85a4f2")]
