+++ SyquenceOfNumbers/SequenceOfNumbers.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35327.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SequenceOfNumbers", "SequenceOfNumbers.csproj", "{9463BDCE-63C3-4A76-8E6F-0AB9AB37D99E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NumberSequence", "..\NumberSequence\NumberSequence.csproj", "{02590964-407D-4D92-BE9E-4DA36EC0963A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SEquenceOfNumbersTest", "..\SEquenceOfNumbersTest\SEquenceOfNumbersTest.csproj", "{680E8F73-2F7E-4F89-B3FA-8C9C2E85A4F2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9463BDCE-63C3-4A76-8E6F-0AB9AB37D99E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9463BDCE-63C3-4A76-8E6F-0AB9AB37D99E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9463BDCE-63C3-4A76-8E6F-0AB9AB37D99E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9463BDCE-63C3-4A76-8E6F-0AB9AB37D99E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{02590964-407D-4D92-BE9E-4DA36EC0963A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02590964-407D-4D92-BE9E-4DA36EC0963A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02590964-407D-4D92-BE9E-4DA36EC0963A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02590964-407D-4D92-BE9E-4DA36EC0963A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{680E8F73-2F7E-4F89-B3FA-8C9C2E85A4F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{680E8F73-2F7E-4F89-B3FA-8C9C2E85A4F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{680E8F73-2F7E-4F89-B3FA-8C9C2E85A4F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{680E8F73-2F7E-4F89-B3FA-8C9C2E85A4F2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4236D5CC-1674-4D23-85BA-1FD1B7EF36FC}
+	EndGlobalSection
+EndGlobal
